Added Generic base class
